First attempt at a driver for the TI OMAP4430 On-die Temperature Sensor
Based on drivers/misc/omap_temp_sensor.c, modified by using
linux-3.16-rc5/drivers/thermal/ti-soc-thermal/ti-bandgap.c, which does
have omap4430 support.

This driver should work on all omap4430 devices, except those where
the SoC pad usbb1_ulpitll_dir is assigned to gpio_86 - in this
case, the omap4430 will internally mux GPIO_22 of the GPIO3 block to
that pad, instead of muxing it to TSHUT. If this is the case, then it
may appear that the TSHUT IRQ is being generated spriously (because
someone else is using that irq...)

To check for this condition, check the appropriate padconf register,
CONTROL_CORE_PAD0_USBB1_ULPITLL_STP_PAD1_USBB1_ULPITLL_DIR
- read physical address 0x4A1000C4
- read the bitfield USBB1_ULPITLL_DIR_MUXMODE by checking bits 18:16.
- If the value read is 0x3, then the usbb1_ulpitll_dir pad is assigned
  to gpio_86 and TSHUT should not be used

In the process, we lose TALERT (configurable alerts for when the
temperature goes above a high setpoint and below a low setpoint), so
we also lose thermal throttle functionality.

We preserve TSHUT functionality even though ti-bandgap.c does not
support it. Hopefully it works. The OMAP4430 TRM says it works.