Include mach-omap2/temp_sensor_device.c functionality in omap4430_temp_sensor
This is necessary because the stock P769 kernel is built without
CONFIG_OMAP_TEMP_SENSOR, so arm/arch/mach-omap2/temp_sensor_device.c is
not compiled in.

This functionality is require to operate the clocks properly so that
when runtime_pm_suspend() and runtime_pm_resume() run, the clocks are
correctly gated.