Reintroduce Throttling Support
Seeing that my P769 hits 92C in stabiliy test while overclocked at
1224 Mhz/1388 mV, and that most PoP LPDDR2 SDRAM memories can only
handle 85C max, we need throttling support.

When the temperature exceeds THROTTLE_HOT, the CPU is throttled (the
highest enabled frequency in the frequency table is disabled). If, one
second later, the temperature still exceeds THROTTLE_HOT, then the CPU
is throttled again, so the top two frequencies are disabled. This
continues. If the temperature is below THROTTLE_HOT but above
THROTTLE_COLD, then the driver will check one second later to see what
the temperature is. If the temperature is below THROTTLE_COLD but above
THROTTLE_COLD - 6 degrees, then the next check occurs in 5 seconds.
Otherwise the temperature is checked every 30 seconds.