Recalculate new throttling thresholds by using the adc/temp table
This allows us to store temperatures in temperatures which can be represented
in the temperature sensor so reading the throttle_temp file provides the actual
thresholds that we are using.

This way, if the user sets the cold threshold and hot threshold to
temperatures which are the same when represented in 8-bit adc
granularity, we can show an error.