Automatic thottling time constants smaller
The commit
0d409157612e Change throttle limits; #define some constants

Fixes the algorithm so that it looks for temperatures strictly greater
than the hot threshold when throttling beyond the first throttle step.
This removes the previous behavior where

The throttle algorithm has throttled the CPU once. The CPU stops getting
hotter, but stays at the same temperature. Without the above mentioned
commit, then the algorithm will throttle again. The above commit changes
this behavior to look for themperatures strictly above the hot throttle
threshold when throttling beyond one step. This allows us to change the
poll interval of the throttling algorithm to one second, because the
possibility of "unneeded throttles" is eliminated.