Redo throttle interval timer settings
Since we need to react to temperatures below the cold threshold
for unthrottling, remove the concept of a "warm" temperature and
only have hot and cold zones for scheduling.

If the temperature is above the cold temperature - 10C, then the next
temperature check occurs in a second, otherwise it will occur in 5
seconds.