omap443x_temp_sensor: Go back to using omap_thermal_throttle()
This is a partial revert of
210f3ea8: omap443x_temp_sensor: Use cpufreq to throttle

When we used cpufreq to throttle, we relied on the interactive CPU
governor to recognize that we changed the cpu min/max frequency limits
and set the cpu speed within the new limits. When the max frequency
decreases (throttle event), then line 1222 of
drivers/cpufreq/cpufreq_interactve.c runs, which causes an immediate
decrease in cpu frequency if the cpu is running above the new maximum.
However, when we raise the maximum (unthrottle event), there is no code
that immediately runs to raise the speed of the cpu. We relied on the
interactive governor to eventually run its heuristics and set the CPU
back to full speed.

However, this does not work correctly because the interactive governor
might not have noticed that the maximum frequency for the processor ever
changed, so it might not apply the new speed until the CPU load changes.

We retain some portions of the partially reverted commit, like the parts
that obtain a reference to the cpufreq driver.