omap_hsmmc_modifier: Modify the suspend backoff algo to backoff more.
The previous commit didn't fix the issue we were having. Probably
because the device "alarm" gets suspended before the mmc hosts, so the
change we made doesn't matter. (But it might matter if a different
device fails to suspend, so keep that change).

Modify the android suspend backoff algorithm to only try twice and then
backoff. It appears that attempting to suspend lots of times in a row
(the defualt is 10 tries before backoff) results the mmc card failure.