suspend_ratelimit: Add new mode with custom wakelock for backing off.
This adds a wake lock and a new operating mode to the suspend_ratelimit driver.
This wake lock can be locked for a set amount of time (default 2.5 seconds) upon
suspend failure.

This is in contrast to the stock suspend_backoff wakelock which would back off
for 10 seconds and waste more power.

Also, enhance the old backoff mode with being able to accept any number of
retries.